requirements.txt
-i https://pypi.org/simple
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
importlib-metadata==4.8.1; python_version < '3.8'
iniconfig==1.1.1
packaging==21.0; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest==6.2.5
pyyaml==5.4.1
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.2; python_version < '3.8'
zipp==3.5.0; python_version >= '3.6'
